% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afstemningsomraader}
\alias{afstemningsomraader}
\title{Data til kort - afstemningsomraader}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1338 rows and 24 columns.
}
\source{
\url{https://api.dataforsyningen.dk/afstemningsomraader?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(afstemningsomraader)
}
\description{
Data til kort - afstemningsomraader
}
\examples{
data(afstemningsomraader)
}
\keyword{datasets}
