% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{DKkommuner}
\alias{DKkommuner}
\title{Dataset med kortdata over DK kommuner}
\format{
A data frame with 99 rows and 12 variables:
\describe{
  \item{dagi_id}{...}
  \item{kode}{...}
  \item{navn}{...}
  \item{regionskode}{...}
  \item{udenforkommuneinddeling}{...}
  \item{regionsnavn}{...}
  \item{ændret}{...}
  \item{geo_ændret}{...}
  \item{geo_version}{...}
  \item{visueltcenter_x}{...}
  \item{visueltcenter_y}{...}
  \item{geometry}{...}
}
}
\usage{
DKkommuner
}
\description{
Dataset med foelgende:
}
\keyword{datasets}
