% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver.R
\name{FT_PARTIER}
\alias{FT_PARTIER}
\title{PARTIER}
\source{
https://www.dr.dk/nyheder/politik/folketingsvalg/resultater & https://www.ft.dk/da/partier/om-politiske-partier/partigruppernes-bogstaver
}
\usage{
FT_PARTIER()
}
\description{
PARTIER
}
