% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver.R
\name{KFUM_PRI}
\alias{KFUM_PRI}
\title{Farvepalet for KFUM-Spejderne}
\usage{
KFUM_PRI()
}
\value{
Farvepalette
}
\description{
Farvepalet for KFUM-Spejderne
}
\examples{
Kontrastfarve = KFUM_KONTRASTFARVER()

}
