% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFUM_themes.R
\name{KFUM_SPEJDERNE_NORMAL}
\alias{KFUM_SPEJDERNE_NORMAL}
\title{GGplot2 themes for KFUM-Spejderne}
\usage{
KFUM_SPEJDERNE_NORMAL(base_family = "", base_size = 11)
}
\arguments{
\item{base_family}{\code{character}, Font typen.}

\item{base_size}{\code{numeric}, Font size.}
}
\value{
A ggplot object.
}
\description{
GGplot2 themes for KFUM-Spejderne
}
\examples{
KFUM_SPEJDERNE_NORMAL(base_family = "", base_size = 11)


}
\keyword{KFUM_SPEJDERNE_NORMAL}
