% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFUM_themes.R
\name{KFUM_SPEJDERNE_TRANSPERENT}
\alias{KFUM_SPEJDERNE_TRANSPERENT}
\title{KFUM_SPEJDERNE_TRANSPERENT}
\usage{
KFUM_SPEJDERNE_TRANSPERENT(base_family = "", base_size = 11)
}
\description{
KFUM_SPEJDERNE_TRANSPERENT
}
