% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kommuner}
\alias{kommuner}
\title{Data til kort - kommuner}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 99 rows and 11 columns.
}
\source{
\url{https://api.dataforsyningen.dk/kommuner?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(kommuner)
}
\description{
Data til kort - kommuner
}
\examples{
data(kommuner)
}
\keyword{datasets}
