% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landekoder}
\alias{landekoder}
\title{ISO standard for landekoder}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 249 rows and 5 columns.
}
\source{
\url{}
}
\usage{
data(landekoder)
}
\description{
ISO standard for landekoder
}
\examples{
data(landekoder)
}
\keyword{datasets}
