% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsdele}
\alias{landsdele}
\title{Data til kort - landsdele}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 11 rows and 10 columns.
}
\source{
\url{https://api.dataforsyningen.dk/landsdele?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(landsdele)
}
\description{
Data til kort - landsdele
}
\examples{
data(landsdele)
}
\keyword{datasets}
