% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opstillingskredse}
\alias{opstillingskredse}
\title{Data til kort - opstillingskredse}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 92 rows and 19 columns.
}
\source{
\url{https://api.dataforsyningen.dk/opstillingskredse?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(opstillingskredse)
}
\description{
Data til kort - opstillingskredse
}
\examples{
data(opstillingskredse)
}
\keyword{datasets}
