% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{PLOTS}
\title{plot.donut}
\usage{
\method{plot}{donut}(df, y, fill, hsize = 3)
}
\arguments{
\item{df}{description}

\item{y}{description}

\item{fill}{description}

\item{hsize}{description}
}
\value{
A ggplot object.
}
\description{
GGplot2 funktioner for plot
}
\details{
\describe{
\item{`plot.donut()`}{.....}
}
}
