% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{politikredse}
\alias{politikredse}
\title{Data til kort - politikredse}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 9 columns.
}
\source{
\url{https://api.dataforsyningen.dk/politikredse?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(politikredse)
}
\description{
Data til kort - politikredse
}
\examples{
data(politikredse)
}
\keyword{datasets}
