% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regioner}
\alias{regioner}
\title{Data til kort - regioner}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 5 rows and 9 columns.
}
\source{
\url{https://api.dataforsyningen.dk/regioner?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(regioner)
}
\description{
Data til kort - regioner
}
\examples{
data(regioner)
}
\keyword{datasets}
