% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retskredse}
\alias{retskredse}
\title{Data til kort - retskredse}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 24 rows and 8 columns.
}
\source{
\url{https://api.dataforsyningen.dk/retskredse?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(retskredse)
}
\description{
Data til kort - retskredse
}
\examples{
data(retskredse)
}
\keyword{datasets}
