% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpakker.R
\name{Rpakker}
\alias{rpakker}
\title{rpakker
Funtkion til at hente og indlæse R-pakker}
\usage{
rpakker(packages_list = c("tidyverse"))
}
\arguments{
\item{packages_list}{liste over R-pakker}
}
\value{
Hent og indlæse R-packages
}
\description{
rpakker
Funtkion til at hente og indlæse R-pakker

rpakker
}
\details{
\describe{
\item{`rpakker()`}{.....}
}
}
\examples{
rpakker(c("tidyverse","sf"))
}
