% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storkredse}
\alias{storkredse}
\title{Data til kort - storkredse}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 12 columns.
}
\source{
\url{https://api.dataforsyningen.dk/storkredse?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(storkredse)
}
\description{
Data til kort - storkredse
}
\examples{
data(storkredse)
}
\keyword{datasets}
