% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valglandsdele}
\alias{valglandsdele}
\title{Data til kort - valglandsdele}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 8 columns.
}
\source{
\url{https://api.dataforsyningen.dk/valglandsdele?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(valglandsdele)
}
\description{
Data til kort - valglandsdele
}
\examples{
data(valglandsdele)
}
\keyword{datasets}
