% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver.R
\name{KFUM_farve}
\title{Farvepalet for KFUM-Spejderne}
\value{
Farvepalette
}
\description{
Farvepalet for KFUM-Spejderne
}
\details{
\describe{
\item{`KFUM_PRI()`}{.....}
\item{`KFUM_SEKUND()`}{.....}
\item{`KFUM_KONTRASTFARVER()`}{.....}
\item{`KFUM_SUPPLEMENT()`}{.....}
\item{`KFUM_GREEN_SCALE()`}{.....}
\item{`KFUM_YELLOW_RED_SCALE()`}{.....}
\item{`FT_PARTIER()`}{.....}
}
}
\examples{
#Return farvepalette
KFUM_PRI()

}
