% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFUM_themes.R
\name{KFUM_tema}
\title{GGplot2 themes for KFUM-Spejderne}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\value{
A ggplot object.
}
\description{
GGplot2 themes for KFUM-Spejderne
}
\details{
\describe{
\item{`KFUM_SPEJDERNE_TRANSPERENT()`}{.....}
\item{`KFUM_SPEJDERNE_NORMAL()`}{.....}
}
}
\examples{
KFUM_SPEJDERNE_NORMAL(base_family = "",base_size = 11)
KFUM_SPEJDERNE_TRANSPERENT(base_family = "",base_size = 11)
GG_TRANSPERENT()

}
