% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot.donut}
\alias{plot.donut}
\title{plot.donut}
\usage{
\method{plot}{donut}(df, y, fill, hsize = 3)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\value{
A ggplot object.
}
\description{
GGplot2 funktioner for plot
}
\details{
\describe{
\item{`plot.donut()`}{.....}
}
}
