% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{Plots}
\title{GGplot2 funktioner for plot}
\usage{
plot_donut(df, y, fill, hsize = 3)
}
\arguments{
\item{df}{description}

\item{y}{description}

\item{fill}{description}

\item{hsize}{description}
}
\value{
Et donut plot.
}
\description{
GGplot2 funktioner for plot

Donut plot
}
\details{
\describe{
   \item{`plot.donut()`}{.....}
}
}
