% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postnumre}
\alias{postnumre}
\title{Data til kort - postnumre}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1089 rows and 10 columns.
}
\source{
\url{https://api.dataforsyningen.dk/postnumre?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(postnumre)
}
\description{
Data til kort - postnumre
}
\examples{
data(postnumre)
}
\keyword{datasets}
