% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpakker.R
\name{rpakker}
\alias{rpakker}
\title{rpakker
Funtkion til at hente og indlæse R-pakker}
\usage{
rpakker(packages_list = c())
}
\arguments{
\item{packages_list}{liste over R-pakker}
}
\value{
Hent og indlæse R-packages
}
\description{
rpakker
Funtkion til at hente og indlæse R-pakker
}
\examples{
rpakker(c("tidyverse","sf"))
}
