% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sogne}
\alias{sogne}
\title{Sognedata til kort}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2129 rows and 9 columns.
}
\source{
\url{https://api.dataforsyningen.dk/sogne?format=geojson}, hentet d. 03/08/2023
}
\usage{
data(sogne)
}
\description{
Sognedata til kort
}
\examples{
data(sogne)
}
\keyword{datasets}
